requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pytz==2023.3
PyYAML==6.0
requests==2.31.0
selenium==3.8.0
urllib3==1.26.6
